connect_orchestrator was refactored
connect #200
