+++ vCenterShell/commands/command_orchestrator.py
@@ -18,6 +18,7 @@ from models.DeployDataHolder import DeployDataHolder
 from models.DriverResponse import DriverResponse, DriverResponseRoot
 from vCenterShell.commands.combine_action import CombineAction
 from vCenterShell.commands.connect_dvswitch import VirtualSwitchConnectCommand
+from vCenterShell.commands.connect_orchestrator import ConnectionCommandOrchestrator
 from vCenterShell.commands.deploy_vm import DeployFromTemplateCommand
 from vCenterShell.commands.destroy_vm import DestroyVirtualMachineCommand
 from vCenterShell.commands.disconnect_dvswitch import VirtualSwitchToMachineDisconnectCommand
@@ -64,8 +65,9 @@ class CommandOrchestrator(object):
         self.command_wrapper = CommandWrapper(logger=getLogger, pv_service=pv_service)
         # Deploy Command
         self.deploy_from_template_command = DeployFromTemplateCommand(deployer=template_deployer)
+
         # Virtual Switch Connect
-        self.virtual_switch_connect_command = \
+        virtual_switch_connect_command = \
             VirtualSwitchConnectCommand(
                 pv_service=pv_service,
                 virtual_switch_to_machine_connector=virtual_switch_to_machine_connector,
@@ -73,6 +75,8 @@ class CommandOrchestrator(object):
                 vlan_spec_factory=VlanSpecFactory(),
                 vlan_id_range_parser=VLanIdRangeParser(),
                 logger=getLogger('VirtualSwitchConnectCommand'))
+        self.connection_orchestrator = ConnectionCommandOrchestrator(self.vc_data_model, virtual_switch_connect_command)
+
         # Virtual Switch Revoke
         self.virtual_switch_disconnect_command = \
             VirtualSwitchToMachineDisconnectCommand(
@@ -91,76 +95,15 @@ class CommandOrchestrator(object):
         # Refresh IP command
         self.refresh_ip_command = RefreshIpCommand(pyvmomi_service=pv_service)
 
-    def _group_actions_by_uuid_and_mode(self, actions):
-        key_to_actions = dict()
-
-        # group by machine and vlan mode and action type
-        for action in actions:
-            vm_uuid = self._get_vm_uuid(action)
-            vlan_mode = action.connectionParams.mode
-            action_type = action.type
-            key = (vm_uuid, vlan_mode, action_type)
-            if key not in key_to_actions:
-                key_to_actions[key] = []
-            key_to_actions[key].append(action)
-
-        # generate new keys
-        return {uuid.uuid4(): action
-                for key, action in key_to_actions}
-
-    def _create_new_action_by_mapping(self, mapping):
-        actions = []
-
-        for key, actions_arr in mapping.items():
-            action = copy.deepcopy(actions_arr[0])
-            action.actionId = key
-            for act in actions_arr[1:]:
-                CombineAction.combine(action, act)
-            actions.append(action)
-        return actions
-
     def connect_bulk(self, context, request):
-        dv_switch_path = self.vc_data_model.default_dvswitch_path
-        dv_switch_name = self.vc_data_model.default_dvswitch_name
-        port_group_path = self.vc_data_model.default_port_group_path
-        default_network = self.vc_data_model.default_network
-
-        holder = DeployDataHolder(jsonpickle.decode(request))
-
-        results = []
         session = self.cs_helper.get_session(context.connectivity.server_address, context.connectivity.admin_auth_token,
                                              context.reservation.domain)
         connection_details = self.cs_helper.get_connection_details(session, context.resource)
 
-        pool = ThreadPool()
-        async_results = []
-
-        mappings = self._group_actions_by_uuid_and_mode(holder.driverRequest.actions, session)
-        unified_actions = self._create_new_action_by_mapping(mappings)
-
-        for key, action in unified_actions.items():
-            vm_uuid = self._get_vm_uuid(action)
-            if action.type == 'setVlan':
-                res = pool.apply_async(self._set_vlan_bulk,
-                                       (action, default_network, dv_switch_name, dv_switch_path,
-                                        port_group_path,
-                                        vm_uuid, connection_details))
-
-            elif action.type == 'removeVlan':
-
-                res = pool.apply_async(self._remove_vlan_bulk,
-                                       (action, default_network, vm_uuid, connection_details))
-
-            if res:
-                async_results.append(res)
-
-        for async_result in async_results:
-            action_results = async_result.get()
-            for action_result in action_results:
-                for unified_action in unified_actions[action_results.actionId]:
-                    copied_action = copy.deepcopy(action_result)
-                    copied_action.actionId = unified_action.actionId
-                    results.append(copied_action)
+        results = self.command_wrapper.execute_command_with_connection(connection_details,
+                                                                       self.connection_orchestrator.connect_bulk,
+                                                                       session,
+                                                                       request)
 
         driver_response = DriverResponse()
         driver_response.actionResults = results
@@ -205,51 +148,6 @@ class CommandOrchestrator(object):
             data_holder)
 
         return set_command_result(result=result, unpicklable=False)
-
-    def connect(self, context, vm_uuid, vlan_id, vlan_spec_type):
-        """
-        Connect Command: connect a vm to a vlan, chooses the first available vnic
-
-        :param context: models.QualiDriverModels.ResourceCommandContext
-        :param str vm_uuid: the uuid id of the  vm
-        :param str vlan_id: the id of the vlan, can be numeric or numeric range
-        :param str vlan_spec_type: the type of the vlan, can only be ('Trunk', 'Access')
-
-        :return str the connection details
-        """
-        # get command parameters from the environment
-        if not vm_uuid:
-            raise ValueError('VM_UUID is missing')
-
-        if not vlan_id:
-            raise ValueError('VLAN_ID is missing')
-
-        if not vlan_spec_type:
-            raise ValueError('VLAN_SPEC_TYPE is missing')
-
-        # load default params
-        vnic_to_network = VmNetworkMapping()
-        vnic_to_network.dv_switch_path = self.vc_data_model.default_dvswitch_path
-        vnic_to_network.dv_switch_name = self.vc_data_model.default_dvswitch_name
-        vnic_to_network.port_group_path = self.vc_data_model.default_port_group_path
-        vnic_to_network.vlan_id = vlan_id
-        vnic_to_network.vlan_spec = vlan_spec_type
-
-        # get connection details
-        session = self.cs_helper.get_session(context.connectivity.server_address, context.connectivity.admin_auth_token,
-                                             context.reservation.domain)
-        connection_details = self.cs_helper.get_connection_details(session, context.resource)
-
-        # execute command
-        connection_results = \
-            self.command_wrapper.execute_command_with_connection(
-                connection_details,
-                self.virtual_switch_connect_command.connect_to_networks,
-                vm_uuid,
-                [vnic_to_network],
-                self.vc_data_model.default_network)
-
-        return set_command_result(result=connection_results, unpicklable=False)
 
     # remote command
     def disconnect_all(self, context, ports):
@@ -399,22 +297,6 @@ class CommandOrchestrator(object):
                                                                    resource_details.fullname)
         return set_command_result(result=res, unpicklable=False)
 
-    def _get_vm_uuid(self, action):
-        vm_uuid_values = [attr.attributeValue for attr in action.customActionAttributes
-                          if attr.attributeName == 'VM_UUID']
-
-        if vm_uuid_values and vm_uuid_values[0]:
-            return vm_uuid_values[0]
-
-        raise ValueError('VM_UUID is missing on action attributes')
-
-    def _get_vnic_name(self, action):
-        vnic_name_values = [attr.attributeValue for attr in action.customActionAttributes
-                            if attr.attributeName == 'Vnic Name']
-        if vnic_name_values:
-            return vnic_name_values[0]
-        return None
-
     def _parse_remote_model(self, context):
         """
         parse the remote resource model and adds its full name
@@ -464,114 +346,3 @@ class CommandOrchestrator(object):
                                                                    resource_details.fullname,
                                                                    self.vc_data_model.default_network)
         return set_command_result(result=res, unpicklable=False)
-
-    def _set_vlan_bulk(self, action, default_network, dv_switch_name, dv_switch_path, port_group_path, vm_uuid,
-                       connection_details):
-        mappings = []
-        results = []
-        for vlan in action.connectionParams.vlanIds:
-            vnic_to_network = VmNetworkMapping()
-            vnic_to_network.dv_switch_path = dv_switch_path
-            vnic_to_network.dv_switch_name = dv_switch_name
-            vnic_to_network.port_group_path = port_group_path
-            vnic_to_network.vlan_id = vlan
-            vnic_to_network.vlan_spec = action.connectionParams.mode
-
-            vnic_name = self._get_vnic_name(action)
-            if vnic_name:
-                vnic_to_network.vnic_name = vnic_name
-
-            mappings.append(vnic_to_network)
-        if mappings:
-            try:
-                connection_results = self.command_wrapper.execute_command_with_connection(connection_details,
-                                                                                          self.virtual_switch_connect_command.connect_to_networks,
-                                                                                          vm_uuid,
-                                                                                          mappings,
-                                                                                          default_network)
-                for connection_result in connection_results:
-                    result = ActionResult()
-                    result.actionId = str(action.actionId)
-                    result.type = str(action.type)
-                    result.infoMessage = 'VLAN successfully set'
-                    result.errorMessage = ''
-                    result.success = True
-                    result.updatedInterface = connection_result.mac_address
-                    results.append(result)
-
-            except Exception as ex:
-                error_result = self._create_failure_result(action, ex)
-                results.append(error_result)
-
-        return results
-
-    @staticmethod
-    def _get_error_message_from_exception(ex):
-        error_message = ''
-        if hasattr(ex, 'msg'):
-            error_message = ex.msg
-        if hasattr(ex, 'faultMessage'):
-            if hasattr(ex.faultMessage, 'message'):
-                error_message += '. ' + ex.faultMessage.message
-        return error_message
-
-    def _remove_vlan_bulk(self, action, default_network, vm_uuid, connection_details):
-        mappings = []
-        results = []
-        interface_attributes = [attr.attributeValue for attr in action.connectorAttributes
-                                if attr.attributeName == 'Interface']
-
-        for interface_attribute in interface_attributes:
-            vm_network_remove_mapping = VmNetworkRemoveMapping()
-            vm_network_remove_mapping.mac_address = interface_attribute
-            vm_network_remove_mapping.vm_uuid = vm_uuid
-            mappings.append(vm_network_remove_mapping)
-
-        if not mappings:
-            error_result = ActionResult()
-            error_result.actionId = str(action.actionId)
-            error_result.type = str(action.type)
-            error_result.infoMessage = str('')
-            error_result.errorMessage = 'Interface attribute is missing on connectorAttributes for removeVlan action'
-            error_result.success = False
-            error_result.updatedInterface = None
-            results = [error_result]
-        else:
-            try:
-                connection_results = self.command_wrapper.execute_command_with_connection(
-                    connection_details,
-                    self.virtual_switch_disconnect_command.disconnect_from_networks,
-                    vm_uuid,
-                    mappings)
-
-                for connection_result in connection_results:
-                    result = self._create_successful_result(action, connection_result)
-                    results.append(result)
-
-            except Exception as ex:
-                error_result = self._create_failure_result(action, ex)
-                results.append(error_result)
-
-        return results
-
-    @staticmethod
-    def _create_successful_result(action, connection_result):
-        result = ActionResult()
-        result.actionId = str(action.actionId)
-        result.type = str(action.type)
-        result.infoMessage = 'VLAN successfully set'
-        result.errorMessage = ''
-        result.success = True
-        result.updatedInterface = connection_result.vnic.macAddress
-        return result
-
-    @staticmethod
-    def _create_failure_result(action, ex):
-        error_result = ActionResult()
-        error_result.actionId = str(action.actionId)
-        error_result.type = str(action.type)
-        error_result.infoMessage = str('')
-        error_result.errorMessage = CommandOrchestrator._get_error_message_from_exception(ex)
-        error_result.success = False
-        error_result.updatedInterface = None
-        return error_result
